requirements.txt
win11toast>=0.35; sys_platform == "win32"
pystray>=0.19
Pillow>=10.0
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
